crcX: Changed lengths to int
Add support for larger buffers, it's hardly worth yielding on 0xff if
you can only hit it ~1 per call on a max-sized buffer.

+++ CRC.h
@@ -74,7 +74,7 @@ uint64_t reverse64(uint64_t in)
 ///////////////////////////////////////////////////////////////////////////////////
 
 // CRC POLYNOME = x8 + x5 + x4 + 1 = 1001 1000 = 0x8C
-uint8_t crc8(const uint8_t *array, uint8_t length, const uint8_t polynome = 0xD5, const uint8_t startmask = 0x00, const uint8_t endmask = 0x00, const bool reverseIn = false, const bool reverseOut = false)
+uint8_t crc8(const uint8_t *array, int length, const uint8_t polynome = 0xD5, const uint8_t startmask = 0x00, const uint8_t endmask = 0x00, const bool reverseIn = false, const bool reverseOut = false)
 {
   uint8_t crc = startmask;
   while (length--) 
@@ -102,7 +102,7 @@ uint8_t crc8(const uint8_t *array, uint8_t length, const uint8_t polynome = 0xD5
 
 
 // CRC POLYNOME = x12 + x3 + x2 + 1 =  0000 1000 0000 1101 = 0x80D
-uint16_t crc12(const uint8_t *array, uint8_t length, const uint16_t polynome = 0x80D, const uint16_t startmask = 0x0000, const uint16_t endmask = 0x0000, const bool reverseIn = false, const bool reverseOut = false)
+uint16_t crc12(const uint8_t *array, int length, const uint16_t polynome = 0x80D, const uint16_t startmask = 0x0000, const uint16_t endmask = 0x0000, const bool reverseIn = false, const bool reverseOut = false)
 {
   uint16_t crc = startmask;
   while (length--) 
@@ -132,7 +132,7 @@ uint16_t crc12(const uint8_t *array, uint8_t length, const uint16_t polynome = 0
 
 
 // CRC POLYNOME = x15 + 1 =  1000 0000 0000 0001 = 0x8001
-uint16_t crc16(const uint8_t *array, uint8_t length, const uint16_t polynome = 0x8001, const uint16_t startmask = 0x0000, const uint16_t endmask = 0x0000, const bool reverseIn = false, const bool reverseOut = false)
+uint16_t crc16(const uint8_t *array, int length, const uint16_t polynome = 0x8001, const uint16_t startmask = 0x0000, const uint16_t endmask = 0x0000, const bool reverseIn = false, const bool reverseOut = false)
 {
   uint16_t crc = startmask;
   while (length--) 
@@ -160,14 +160,14 @@ uint16_t crc16(const uint8_t *array, uint8_t length, const uint16_t polynome = 0
 
 
 // CRC-CCITT POLYNOME = x13 + X5 + 1 =  0001 0000 0010 0001 = 0x1021
-uint16_t crc16_CCITT(uint8_t *array, uint8_t length)
+uint16_t crc16_CCITT(uint8_t *array, int length)
 {
   return crc16(array, length, 0x1021, 0xFFFF);
 }
 
 
 // CRC-32 POLYNOME =  x32 + ..... + 1
-uint32_t crc32(const uint8_t *array, uint8_t length, const uint32_t polynome = 0x04C11DB7, const uint32_t startmask = 0, const uint32_t endmask = 0, const bool reverseIn = false, const bool reverseOut = false)
+uint32_t crc32(const uint8_t *array, int length, const uint32_t polynome = 0x04C11DB7, const uint32_t startmask = 0, const uint32_t endmask = 0, const bool reverseIn = false, const bool reverseOut = false)
 {
   uint32_t crc = startmask;
   while (length--) 
@@ -195,7 +195,7 @@ uint32_t crc32(const uint8_t *array, uint8_t length, const uint32_t polynome = 0
 
 
 // CRC-CCITT POLYNOME =  x64 + ..... + 1
-uint64_t crc64(const uint8_t *array, uint8_t length, const uint64_t polynome, const uint64_t startmask, const uint64_t endmask, const bool reverseIn, const bool reverseOut)
+uint64_t crc64(const uint8_t *array, int length, const uint64_t polynome, const uint64_t startmask, const uint64_t endmask, const bool reverseIn, const bool reverseOut)
 {
   uint64_t crc = startmask;
   while (length--)
