+++ CRC32.cpp
@@ -44,7 +44,7 @@ void CRC32::add(uint8_t value)
 }
 
 
-void CRC32::add(const uint8_t * array, uint8_t length)
+void CRC32::add(const uint8_t * array, int length)
 {
   while (length--)
     add(*array++);
